Move WitnessVersion to primitives
As part of moving Address to addresses and the script functionality to
primitives, WitnessVersion will both also need to be available to both.

Move WitnessVersion and associated errors to primitives, re-exporting
from bitcoin.
